backend/requirements.txt
pymysql
sqlalchemy
fastapi
uvicorn
dotenv
alembic
bcrypt
PyJWT
pandas
numpy
scikit-learn
matplotlib
imbalanced-learn
pymongo>=4.6.0
dnspython>=2.4.0
